# docs/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "ThueXe",
    "version": "1.0"
  },
  "paths": {
...
    "/me/bank-account": {
      "put": {
        "tags": [
          "Me"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatTaiKhoanRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatTaiKhoanRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CapNhatTaiKhoanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MeResponse"
                }
              }
            }
          }
        }
      }
    },
    "/me/documents": {
      "post": {
        "tags": [
          "Me"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDocumentsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDocumentsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitDocumentsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CapNhatTaiKhoanRequest": {
        "type": "object",
        "properties": {
          "bankAccount": {
            "type": "string",
            "nullable": true
          },
          "bankName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "MeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "fullName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "isOwner": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "nullable": true
          },
          "documentStatus": {
            "type": "string",
            "nullable": true
          },
          "bankAccount": {
            "type": "string",
            "nullable": true
          },
          "bankName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "SubmitDocumentsRequest": {
        "type": "object",
        "properties": {
          "cccdNo": {
            "type": "string",
            "nullable": true
          },
          "gplxNo": {
            "type": "string",
            "nullable": true
          },
          "gplxClass": {
            "type": "string",
            "nullable": true
          },
          "gplxExpiry": {
            "type": "string",
            "format": "date",
            "nullable": true
          },
          "frontUrl": {
            "type": "string",
            "nullable": true
          },
          "backUrl": {
            "type": "string",
            "nullable": true
          },
          "selfieUrl": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "description": "Nhập JWT token",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearer": [ ]
    }
  ],
...
}